Refactor s2n_record_read/write.c, move functions to cipher type files and minor style changes
Currently s2n_record_read.c has function s2n_record_parse for message decryption
and s2n_record_write.c has function s2n_record_write for encryption. Both functions
are very long and have lots of logics based on different cipher types. Extracting those
functions into files based on cipher types(s2n_cbc_record.c, s2n_composite_record.c,
s2n_stream_record.c and s2n_aead_record.c) allows users to more easily understand the code,
and adds flexibility on potential changes in specific cipher type.
