+++ tests/unit/s2n_aead_chacha20_poly1305_test.c
@@ -79,7 +79,7 @@ int main(int argc, char **argv)
         conn->client_protocol_version = S2N_TLS12;
         conn->actual_protocol_version = S2N_TLS12;
         EXPECT_SUCCESS(setup_server_keys(conn, &chacha20_poly1305_key));
-        EXPECT_SUCCESS(bytes_written = s2n_record_write(conn, TLS_APPLICATION_DATA, &in));
+        EXPECT_SUCCESS(bytes_written = conn->secure.cipher_suite->record_alg->record_write(conn, TLS_APPLICATION_DATA, &in));
 
         static const int overhead = S2N_TLS_CHACHA20_POLY1305_EXPLICIT_IV_LEN   /* Should be 0 */
             + S2N_TLS_GCM_TAG_LEN; /* TAG */
@@ -114,7 +114,7 @@ int main(int argc, char **argv)
         uint8_t content_type;
         uint16_t fragment_length;
         EXPECT_SUCCESS(s2n_record_header_parse(conn, &content_type, &fragment_length));
-        EXPECT_SUCCESS(s2n_record_parse(conn));
+        EXPECT_SUCCESS(conn->secure.cipher_suite->record_alg->record_parse(conn));
         EXPECT_EQUAL(content_type, TLS_APPLICATION_DATA);
         EXPECT_EQUAL(fragment_length, predicted_length);
 
@@ -127,7 +127,7 @@ int main(int argc, char **argv)
         conn->client_protocol_version = S2N_TLS12;
         conn->actual_protocol_version = S2N_TLS12;
         EXPECT_SUCCESS(setup_server_keys(conn, &chacha20_poly1305_key));
-        EXPECT_SUCCESS(s2n_record_write(conn, TLS_APPLICATION_DATA, &in));
+        EXPECT_SUCCESS(conn->secure.cipher_suite->record_alg->record_write(conn, TLS_APPLICATION_DATA, &in));
 
         /* Now lets corrupt some data and ensure the tests pass */
         /* Copy the encrypted out data to the in data */
@@ -139,7 +139,7 @@ int main(int argc, char **argv)
         /* Tamper the protocol version in the header, and ensure decryption fails, as we use this in the AAD */
         conn->in.blob.data[2] = 2;
         EXPECT_SUCCESS(s2n_record_header_parse(conn, &content_type, &fragment_length));
-        EXPECT_FAILURE(s2n_record_parse(conn));
+        EXPECT_FAILURE(conn->secure.cipher_suite->record_alg->record_parse(conn));
         EXPECT_EQUAL(content_type, TLS_APPLICATION_DATA);
 
         EXPECT_SUCCESS(s2n_stuffer_wipe(&conn->header_in));
@@ -154,7 +154,7 @@ int main(int argc, char **argv)
             conn->client_protocol_version = S2N_TLS12;
             conn->actual_protocol_version = S2N_TLS12;
             EXPECT_SUCCESS(setup_server_keys(conn, &chacha20_poly1305_key));
-            EXPECT_SUCCESS(s2n_record_write(conn, TLS_APPLICATION_DATA, &in));
+            EXPECT_SUCCESS(conn->secure.cipher_suite->record_alg->record_write(conn, TLS_APPLICATION_DATA, &in));
 
             /* Copy the encrypted out data to the in data */
             EXPECT_SUCCESS(s2n_stuffer_wipe(&conn->in));
@@ -163,7 +163,7 @@ int main(int argc, char **argv)
             EXPECT_SUCCESS(s2n_stuffer_copy(&conn->out, &conn->in, s2n_stuffer_data_available(&conn->out)));
             conn->in.blob.data[s2n_stuffer_data_available(&conn->in) - j - 1] ++;
             EXPECT_SUCCESS(s2n_record_header_parse(conn, &content_type, &fragment_length));
-            EXPECT_FAILURE(s2n_record_parse(conn));
+            EXPECT_FAILURE(conn->secure.cipher_suite->record_alg->record_parse(conn));
             EXPECT_EQUAL(content_type, TLS_APPLICATION_DATA);
 
             EXPECT_SUCCESS(s2n_stuffer_wipe(&conn->header_in));
@@ -179,7 +179,7 @@ int main(int argc, char **argv)
             conn->client_protocol_version = S2N_TLS12;
             conn->actual_protocol_version = S2N_TLS12;
             EXPECT_SUCCESS(setup_server_keys(conn, &chacha20_poly1305_key));
-            EXPECT_SUCCESS(s2n_record_write(conn, TLS_APPLICATION_DATA, &in));
+            EXPECT_SUCCESS(conn->secure.cipher_suite->record_alg->record_write(conn, TLS_APPLICATION_DATA, &in));
 
             /* Copy the encrypted out data to the in data */
             EXPECT_SUCCESS(s2n_stuffer_wipe(&conn->in));
@@ -188,7 +188,7 @@ int main(int argc, char **argv)
             EXPECT_SUCCESS(s2n_stuffer_copy(&conn->out, &conn->in, s2n_stuffer_data_available(&conn->out)));
             conn->in.blob.data[S2N_TLS_GCM_IV_LEN + j]++;
             EXPECT_SUCCESS(s2n_record_header_parse(conn, &content_type, &fragment_length));
-            EXPECT_FAILURE(s2n_record_parse(conn));
+            EXPECT_FAILURE(conn->secure.cipher_suite->record_alg->record_parse(conn));
             EXPECT_EQUAL(content_type, TLS_APPLICATION_DATA);
 
             EXPECT_SUCCESS(s2n_stuffer_wipe(&conn->header_in));
