+++ tls/s2n_cipher_suites.h
@@ -47,6 +47,8 @@ struct s2n_record_algorithm {
     const struct s2n_cipher *cipher;
     s2n_hmac_algorithm hmac_alg;
     uint32_t flags;
+    int (*record_write) (struct s2n_connection *conn, uint8_t content_type, struct s2n_blob *in);
+    int (*record_parse) (struct s2n_connection *conn);
 };
 
 /* Verbose names to avoid confusion with s2n_cipher. Exposed for unit tests */
