+++ tls/s2n_handshake_io.c
@@ -270,7 +270,11 @@ static int handshake_write_io(struct s2n_connection *conn)
         notnull_check(out.data);
 
         /* Make the actual record */
-        GUARD(s2n_record_write(conn, record_type, &out));
+        if (conn->mode == S2N_CLIENT) {
+            GUARD(conn->client->cipher_suite->record_alg->record_write(conn, record_type, &out));
+        } else {
+            GUARD(conn->server->cipher_suite->record_alg->record_write(conn, record_type, &out));
+        }
 
         /* MD5 and SHA sum the handshake data too */
         if (record_type == TLS_HANDSHAKE) {
